+++ CHANGES.rst
@@ -7,6 +7,8 @@ Changelog
 
 - Added `PloneGroupUsersGroupsColumn`, a column that displays suffixed groups
   and users, to be called on dashboard displaying organizations.
+  The groups and users are rendered by the `@@display-group-users` view
+  that may also be used outside.
   [gbastien]
 - Added possibility to disable a `function` (`enabled=True` by default),
   this is useful to avoid deleting a `function` and re-adding it after,